Remove the v1 example because there was a tag problem on pongo2
Any tag before v4 on the pongo2 repository and it says something along the lines of that tag or commit not existing in the repo.

Since v1 is now old, it's easier just to drop it, tried figuring it out but gave up in the end.

Also update LICENSE to contain the text "and contributors" to credit contributors.

Also remove .gitattributes because I never approved this, AND it says "don't lint html and css" which I disagree with, I see problems in the HTML already (unclosed li tag).

In the next and final phase I will be updating the example template and fixing the linting problems I'm seeing in GoLand.