Bugfix: Update Context with TemplateSet.Globals is redundant and modifies data map
This line is redundant, I proved it with a test. I'll commit a test suite later, but I have proven this line is NOT needed:

data.(pongo2.Context).Update(p.Options.TemplateSet.Globals)

In fact having that line there just modifies the data map after it leaves the Instance method. We definitely don't want that.

Globals are automatically sent to templates, you don't need to do this yourself.

Closes #8