Set TemplateSet.Debug in New
When we create a TemplateSet for the TemplateDir in New, we can set TemplateSet.Debug to true if GIN_MODE=debug, otherwise set TemplateSet.Debug to false

This means we can simplify the Instance function and always call TemplateSet.FromCache, it will call .FromFile if TemplateSet.Debug is true.

To test this, I removed the gin.SetMode(gin.DebugMode) from the example, or you can't test it.

    GIN_MODE=release ./example
    ./example

It seems that Gin defaults to Debug mode being on which is a bit strange.

If the user has passed in a custom TemplateSet, then we would rather not set Debug on it and let the caller do it.